Formating login and register page login and register function should have their own file
Splitted the login and register function into their own file

Note
