+++ register.php
@@ -1,38 +1,5 @@
 <?php
-require_once("includes/config.php");
-require_once("includes/classes/FormSanitizer.php");
-require_once("includes/classes/Constants.php");
-require_once("includes/classes/Account.php");
-    
-    $account = new Account($conn);
-
-    if(isset($_POST["submitButton"]))
-    {
-        $firstName = FormSanitizer::sanitizeFormString($_POST["firstName"]);
-        $lastName = FormSanitizer::sanitizeFormString($_POST["lastName"]);
-        $userName = FormSanitizer::sanitizeFormUsername($_POST["username"]);
-        $email = FormSanitizer::sanitizeFormEmail($_POST["email"]);
-        $email2 = FormSanitizer::sanitizeFormEmail($_POST["email2"]);
-        $password = FormSanitizer::sanitizeFormPassword($_POST["password"]);
-        $password2 = FormSanitizer::sanitizeFormPassword($_POST["password"]);
-
-        // if the input is successful, register will try to insert user details
-        $success = $account->register($firstName, $lastName, $userName, $email, $email2, $password, $password2);
-        // if the insertion is success, register will hit the if block below
-        if($success)
-        {
-            $_SESSION["userLoggedIn"] = $username;
-            header("Location: index.php");
-        }
-    }
-    
-    function getInputValue($name)
-    {
-        if(isset($_POST[$name]))
-        {
-            echo $_POST[$name];
-        }
-    }
+require_once("includes/processForm/registerProcess.php");
 ?>
 <!DOCTYPE html>
 <html lang="en">
